Update iOS13 branching logic for card scanning
