+++ AdyenCard/Components/Card/CardScannerController.swift
@@ -28,6 +28,7 @@ internal protocol CardScannerControlling: CardScannerAvailability {
 #if canImport(AdyenCardScanner)
     import AdyenCardScanner
 
+    @available(iOS 13.0, *)
     private struct CardScannerAvailabilityWrapper: CardScannerAvailability {
         var isScannerAvailable: Bool {
             AdyenCardScanner.CardScanner.isAvailable
@@ -52,6 +53,7 @@ internal protocol CardScannerControlling: CardScannerAvailability {
         }
     }
 
+    @available(iOS 13.0, *)
     internal struct CardScannerProviderWrapper: CardScannerProviding {
         internal func createCardScanner(completion: @escaping (Result<CardScannerCardDetails, Error>) -> Void) -> UIViewController? {
 
@@ -69,6 +71,7 @@ internal protocol CardScannerControlling: CardScannerAvailability {
         }
     }
 
+    @available(iOS 13.0, *)
     internal final class CardScannerController: CardScannerControlling {
         internal enum CardScannerError: Error {
             case scanningError
@@ -140,8 +143,6 @@ internal protocol CardScannerControlling: CardScannerAvailability {
         }
 
         private func makeNavigationController() -> UINavigationController {
-            guard #available(iOS 13.0, *) else { return UINavigationController() }
-
             let appearance = UINavigationBarAppearance()
             appearance.configureWithDefaultBackground()
 
@@ -176,32 +177,6 @@ internal protocol CardScannerControlling: CardScannerAvailability {
 
 #else // canImport(AdyenCardScanner)
 
-    internal final class CardScannerController: CardScannerControlling {
-        internal var isScannerAvailable: Bool { false }
-        internal var onScanComplete: ((Result<CardScannerCardDetails, any Error>) -> Void)?
-        internal var title: String?
-        internal func openCardScanner() {}
-
-        internal init(
-            presenter: UIViewController,
-            availabilityProvider: CardScannerAvailability = DummyCardScannerAvailability(),
-            cardScannerProvider: CardScannerProviding = DummyCardScannerProvider(),
-            analyticsHandler: CardScannerAnalyticsHandler?
-        ) {}
-
-        // MARK: - Helpers
-
-        internal struct DummyCardScannerAvailability: CardScannerAvailability {
-            internal var isScannerAvailable: Bool { false }
-        }
-
-        internal struct DummyCardScannerProvider: CardScannerProviding {
-            internal func createCardScanner(
-                completion: @escaping (Result<CardScannerCardDetails, any Error>) -> Void
-            ) -> UIViewController? {
-                UIViewController()
-            }
-        }
-    }
+    typealias CardScannerController = DummyCardScannerController
 
 #endif // canImport(AdyenCardScanner)
